Move EditorProps to types.ts
